Added Category column to euclid, angle, and MSD sheets. Added Final MSD column to summary sheet. For both straightness and convex hull, the time-corrected version is the only one shown. Now also outputting a second PCA figure that uses PCs 2,3,4.

+++ main.py
@@ -502,7 +502,7 @@ def run_migrate(*vals):
     args = [
         parent_id, time_for, x_for, y_for, z_for,
         float(timelapse), float(arrest_limit), int(moving),
-        int(contact_length), float(arrested), int(tau),
+        float(contact_length), float(arrested), int(tau),
         formatting_options, savefile,
         segments_file, category_file,
         parent_id2, category_col_name,
